Data Science/Crop Disease Detection/requirements.txt
numpy
streamlit
tensorflow
flask


fastapi
uvicorn
pillow
python-multipart

